Some naming changes
